+++ src/rheolwyr/database.py
@@ -1,5 +1,5 @@
 # Rheolwyr - Linux Text Expander
-# Copyright (C) 2026 Chuck Talk <cwtalk1@gmail.com>
+# Copyright (C) 2026 Chuck Talk <chuck@nordheim.online>
 # Licensed under GPLv3 or later
 
 import os